rename cssSnapshots.ts => DtsSnapshotCreator
